refactor: 카테고리 서비스 칼질중
getAllCategories 컨트롤러에서 직접 수행
