+++ src/main/java/com/meoguri/linkocean/controller/bookmark/CategoryController.java
@@ -1,24 +1,19 @@
-package com.meoguri.linkocean.controller.category;
+package com.meoguri.linkocean.controller.bookmark;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.meoguri.linkocean.controller.common.SliceResponse;
-import com.meoguri.linkocean.domain.bookmark.service.CategoryService;
+import com.meoguri.linkocean.domain.bookmark.entity.Bookmark;
 
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
 @RequestMapping("/api/v1/categories")
 @RestController
 public class CategoryController {
 
-	private final CategoryService categoryService;
-
 	@GetMapping
 	public SliceResponse<String> getAllCategories() {
 
-		return SliceResponse.of("categories", categoryService.getAllCategories());
+		return SliceResponse.of("categories", Bookmark.Category.getKoreanNames());
 	}
 }
